app/requirements.txt
pydantic~=2.12.5
httpx~=0.28.1
uvicorn~=0.40.0
fastapi~=0.128.0
pydantic-settings~=2.12.0
PyYAML~=6.0.3